Add Agendas

+++ app/Http/Controllers/Admin/OrganisationController.php
@@ -8,6 +8,7 @@ use App\Models\Organisation;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 use Inertia\Response;
@@ -16,6 +17,7 @@ class OrganisationController extends Controller
 {
     public function index(): Response
     {
+        Gate::authorize('viewAny', Organisation::class);
         $organisations = Organisation::all();
 
         return Inertia::render('dashboard/organisations/Index', ['organisations' => $organisations]);
@@ -23,6 +25,7 @@ class OrganisationController extends Controller
 
     public function show(Organisation $organisation): Response
     {
+        Gate::authorize('view', Organisation::class);
         $organisation->load('users');
 
         return Inertia::render('dashboard/organisations/Show', ['organisation' => $organisation]);
@@ -30,7 +33,7 @@ class OrganisationController extends Controller
 
     public function update(UpdateOrganisationRequest $request, Organisation $organisation): RedirectResponse
     {
-
+        Gate::authorize('update', Organisation::class);
         if ($request->hasFile('image')) {
             if ($organisation->image_path && Storage::disk('hetzner')->exists($organisation->image_path)) {
                 Storage::disk('hetzner')->delete($organisation->image_path);
@@ -50,6 +53,7 @@ class OrganisationController extends Controller
 
     public function detachUser(Request $request, Organisation $organisation, User $user): RedirectResponse
     {
+        Gate::authorize('update', Organisation::class);
         $organisation->users()->detach($user->id);
 
         return redirect()->route('admin.organisations.show', $organisation->slug);
